+++ packages/blocks-core/js/libs/wordpress/term-name/test/test-content.js
@@ -1,0 +1,11 @@
+export const testContent = ` 
+<!-- wp:terms-query -->
+<div class="wp-block-terms-query"><!-- wp:term-template -->
+<!-- wp:group {"layout":{"type":"flex","flexWrap":"nowrap"}} -->
+<div class="wp-block-group"><!-- wp:term-name {"isLink":true} /-->
+
+<!-- wp:term-count /--></div>
+<!-- /wp:group -->
+<!-- /wp:term-template --></div>
+<!-- /wp:terms-query -->
+`;