+++ packages/blocks-core/CHANGELOG.md
@@ -35,6 +35,7 @@
 - Added E2E tests to check functionality and support of `Post Comments Count` block.
 - Added E2E tests to check functionality and support of `Post Comments Link` block.
 - Added E2E tests to check functionality and support of `Term Template` block.
+- Added E2E tests to check functionality and support of `Term Query → Term Name` block.
 
 ## 1.4.0 (2025-05-19)